feat: Add database new operations on generated classes.
