+++ tools/serverpod_cli/test/generator/dart/server_code_generator/database_repository_class_test.dart
@@ -12,6 +12,7 @@ const generator = DartServerCodeGenerator();
 
 void main() {
   var testClassName = 'Example';
+  var repositoryClassName = '${testClassName}Repository';
   var testClassFileName = 'example';
   var expectedFilePath =
       path.join('lib', 'src', 'generated', '$testClassFileName.dart');
@@ -32,422 +33,474 @@ void main() {
 
     var compilationUnit = parseString(content: codeMap[expectedFilePath]!).unit;
 
-    test('then a class named ${testClassName}AttachRepository is NOT generated',
-        () {
-      expect(
-        CompilationUnitHelpers.hasClassDeclaration(
-          compilationUnit,
-          name: '${testClassName}AttachRepository',
-        ),
-        isFalse,
-        reason:
-            'The class ${testClassName}AttachRepository was found but was expected to not exist.',
-      );
-    });
-
-    test('then a class named ${testClassName}DetachRepository is NOT generated',
-        () {
-      expect(
-        CompilationUnitHelpers.hasClassDeclaration(
-          compilationUnit,
-          name: '${testClassName}DetachRepository',
-        ),
-        isFalse,
-        reason:
-            'The class ${testClassName}DetachRepository was found but was expected to not exist.',
-      );
-    });
-  });
-  group(
-      'Given a class with table name and object relation field when generating code',
-      () {
-    var entities = [
-      ClassDefinitionBuilder()
-          .withClassName(testClassName)
-          .withFileName(testClassFileName)
-          .withTableName('example_table')
-          .withObjectRelationField(
-            'company',
-            'Company',
-            'company',
-            nullableRelation: true,
-          )
-          .withObjectRelationField(
-            'address',
-            'Address',
-            'address',
-            foreignKeyFieldName: 'companyAddressId',
-          )
-          .build()
-    ];
-
-    var codeMap = generator.generateSerializableEntitiesCode(
-      entities: entities,
-      config: config,
+    var baseClass = CompilationUnitHelpers.tryFindClassDeclaration(
+      compilationUnit,
+      name: testClassName,
     );
 
-    var compilationUnit = parseString(content: codeMap[expectedFilePath]!).unit;
+    test('then a static db field is generated on the base class.', () {
+      var dbField = CompilationUnitHelpers.tryFindFieldDeclaration(
+        baseClass!,
+        name: 'db',
+      );
+
+      expect(
+        dbField?.fields.toSource(),
+        'final db = $repositoryClassName._()',
+      );
+    });
 
     var repositoryClass = CompilationUnitHelpers.tryFindClassDeclaration(
       compilationUnit,
-      name: '${testClassName}Repository',
+      name: repositoryClassName,
     );
 
-    group('then the class name ${testClassName}Repository', () {
-      test('has a final attach field', () {
-        var field = CompilationUnitHelpers.tryFindFieldDeclaration(
-          repositoryClass!,
-          name: 'attach',
-        );
-
-        expect(
-          field?.toSource(),
-          'final attach = const ${testClassName}AttachRepository._();',
-          reason: 'Missing static instance field.',
-        );
-      });
-
-      test('has a final detach field', () {
-        var field = CompilationUnitHelpers.tryFindFieldDeclaration(
-          repositoryClass!,
-          name: 'detach',
-        );
-
-        expect(
-          field?.toSource(),
-          'final detach = const ${testClassName}DetachRepository._();',
-          reason: 'Missing static instance field.',
-        );
-      });
-    }, skip: repositoryClass == null);
-
-    test('then a class named ${testClassName}AttachRepository is generated',
-        () {
+    test('then the class name $repositoryClassName is generated', () {
       expect(
-        CompilationUnitHelpers.hasClassDeclaration(
-          compilationUnit,
-          name: '${testClassName}AttachRepository',
-        ),
-        isTrue,
-        reason:
-            'Expected the class ${testClassName}AttachRepository to be generated.',
-      );
-    });
-
-    var repositoryAttachClass = CompilationUnitHelpers.tryFindClassDeclaration(
-      compilationUnit,
-      name: '${testClassName}AttachRepository',
-    );
-
-    group('then the ${testClassName}AttachRepository', () {
-      test('has a private constructor', () {
-        var constructor = CompilationUnitHelpers.tryFindConstructorDeclaration(
-          repositoryAttachClass!,
-          name: '_',
-        );
-        expect(
-          constructor?.toSource(),
-          'const ${testClassName}AttachRepository._();',
-          reason: 'Missing private constructor.',
-        );
-      });
-
-      var companyMethod = CompilationUnitHelpers.tryFindMethodDeclaration(
-        repositoryAttachClass!,
-        name: 'company',
-      );
-
-      test('has a company method defined.', () {
-        expect(companyMethod, isNotNull, reason: 'Missing company method.');
-      });
-
-      test(
-          'company method has the input params of session, example and company',
-          () {
-        expect(
-          companyMethod?.parameters?.toSource(),
-          matches(
-            r'(_i\d.Session session, Example example, Company company)',
-          ),
-        );
-      }, skip: companyMethod == null);
-
-      var addressMethod = CompilationUnitHelpers.tryFindMethodDeclaration(
-        repositoryAttachClass,
-        name: 'address',
-      );
-
-      test('has a address method defined.', () {
-        expect(addressMethod, isNotNull, reason: 'Missing address method.');
-      });
-
-      test(
-          'the address method has the input params of session, example and address',
-          () {
-        expect(
-          addressMethod?.parameters?.toSource(),
-          matches(
-            r'(_i\d.Session session, Example example, Address address)',
-          ),
-        );
-      }, skip: addressMethod == null);
-
-      test('has no method for the id field', () {
-        expect(
-          CompilationUnitHelpers.hasMethodDeclaration(
-            repositoryAttachClass,
-            name: 'id',
-          ),
-          isFalse,
-        );
-      });
-
-      test('has no method for the companyId field', () {
-        expect(
-          CompilationUnitHelpers.hasMethodDeclaration(
-            repositoryAttachClass,
-            name: 'companyId',
-          ),
-          isFalse,
-        );
-      });
-
-      test('has no method for the companyAddressId field', () {
-        expect(
-          CompilationUnitHelpers.hasMethodDeclaration(
-            repositoryAttachClass,
-            name: 'companyAddressId',
-          ),
-          isFalse,
-        );
-      });
-    }, skip: repositoryClass == null);
-
-    test('then a class named ${testClassName}DetachRepository is generated',
-        () {
-      expect(
-        CompilationUnitHelpers.hasClassDeclaration(
-          compilationUnit,
-          name: '${testClassName}DetachRepository',
-        ),
-        isTrue,
-        reason:
-            'Expected the class ${testClassName}DetachRepository to be generated.',
-      );
-    });
-
-    var repositoryDetachClass = CompilationUnitHelpers.tryFindClassDeclaration(
-      compilationUnit,
-      name: '${testClassName}DetachRepository',
-    );
-    group('then the ${testClassName}DetachRepository', () {
-      test('has a private constructor', () {
-        var constructor = CompilationUnitHelpers.tryFindConstructorDeclaration(
-          repositoryDetachClass!,
-          name: '_',
-        );
-        expect(
-          constructor?.toSource(),
-          'const ${testClassName}DetachRepository._();',
-          reason: 'Missing private constructor.',
-        );
-      });
-
-      var companyMethod = CompilationUnitHelpers.tryFindMethodDeclaration(
-        repositoryDetachClass!,
-        name: 'company',
-      );
-
-      test('has a company method defined.', () {
-        expect(companyMethod, isNotNull, reason: 'Missing company method.');
-      });
-
-      test('company method has the input params of session, example', () {
-        expect(
-          companyMethod?.parameters?.toSource(),
-          '(_i1.Session session, Example example)',
-        );
-      }, skip: companyMethod == null);
-
-      var addressMethod = CompilationUnitHelpers.tryFindMethodDeclaration(
-        repositoryDetachClass,
-        name: 'address',
-      );
-
-      test('has NOT an address method defined for none nullable relation.', () {
-        expect(addressMethod, isNull, reason: 'Missing address method.');
-      });
-
-      test('has no method for the id field', () {
-        expect(
-          CompilationUnitHelpers.hasMethodDeclaration(
-            repositoryDetachClass,
-            name: 'id',
-          ),
-          isFalse,
-        );
-      });
-
-      test('has no method for the companyId field', () {
-        expect(
-          CompilationUnitHelpers.hasMethodDeclaration(
-            repositoryDetachClass,
-            name: 'companyId',
-          ),
-          isFalse,
-        );
-      });
-
-      test('has no method for the companyAddressId field', () {
-        expect(
-          CompilationUnitHelpers.hasMethodDeclaration(
-            repositoryDetachClass,
-            name: 'companyAddressId',
-          ),
-          isFalse,
-        );
-      });
-    }, skip: repositoryClass == null);
-  });
-
-  group(
-      'Given a class with table name and object relation field when generating code',
-      () {
-    var entities = [
-      ClassDefinitionBuilder()
-          .withClassName(testClassName)
-          .withFileName(testClassFileName)
-          .withTableName('example')
-          .withObjectRelationField('company', 'Company', 'company')
-          .withObjectRelationFieldNoForeignKey(
-            'address',
-            'Address',
-            'address',
-          )
-          .build()
-    ];
-
-    var codeMap = generator.generateSerializableEntitiesCode(
-      entities: entities,
-      config: config,
-    );
-
-    var compilationUnit = parseString(content: codeMap[expectedFilePath]!).unit;
-
-    var repositoryDetachClass = CompilationUnitHelpers.tryFindClassDeclaration(
-      compilationUnit,
-      name: '${testClassName}DetachRepository',
-    );
-
-    test('then no detach repository is generated', () {
-      expect(
-        repositoryDetachClass,
-        isNull,
-        reason:
-            'The class ${testClassName}DetachRepository was found but was expected to not exist.',
-      );
-    });
-  });
-
-  group(
-      'Given a class with table name and object relation field when generating code',
-      () {
-    var entities = [
-      ClassDefinitionBuilder()
-          .withClassName(testClassName)
-          .withFileName(testClassFileName)
-          .withTableName('example')
-          .withObjectRelationField(
-            'company',
-            'Company',
-            'company',
-            nullableRelation: true,
-          )
-          .withObjectRelationFieldNoForeignKey(
-            'address',
-            'Address',
-            'address',
-          )
-          .build()
-    ];
-
-    var codeMap = generator.generateSerializableEntitiesCode(
-      entities: entities,
-      config: config,
-    );
-
-    var compilationUnit = parseString(content: codeMap[expectedFilePath]!).unit;
-
-    var repositoryDetachClass = CompilationUnitHelpers.tryFindClassDeclaration(
-      compilationUnit,
-      name: '${testClassName}DetachRepository',
-    );
-
-    group(
-        'then the address method is not generated for none nullable relation.',
-        () {
-      var addressMethod = CompilationUnitHelpers.tryFindMethodDeclaration(
-        repositoryDetachClass!,
-        name: 'address',
-      );
-
-      test('', () {
-        expect(addressMethod, isNull, reason: 'Missing address method.');
-      });
-    });
-  });
-
-  group(
-      'Given a class with a self relation where the field has the same name as the class',
-      () {
-    var tableName = 'example_table';
-    var entities = [
-      ClassDefinitionBuilder()
-          .withFileName(testClassFileName)
-          .withTableName(tableName)
-          .withObjectRelationField(
-            'example',
-            testClassName,
-            tableName,
-            nullableRelation: true,
-          )
-          .build()
-    ];
-
-    var codeMap = generator.generateSerializableEntitiesCode(
-      entities: entities,
-      config: config,
-    );
-
-    var compilationUnit = parseString(content: codeMap[expectedFilePath]!).unit;
-
-    var attachRepository = CompilationUnitHelpers.tryFindClassDeclaration(
-      compilationUnit,
-      name: '${testClassName}AttachRepository',
-    );
-
-    group('then the attach method for example', () {
-      var method = CompilationUnitHelpers.tryFindMethodDeclaration(
-        attachRepository!,
-        name: 'example',
-      );
-
-      test('has the secondary input param named "nestedExample" ', () {
-        expect(
-          method?.parameters?.toSource(),
-          matches(
-            r'(_i\d.Session session, Example example, Example nestedExample)',
-          ),
-        );
-      });
-    });
-
-    test('then the class name ${testClassName}Repository is generated', () {
-      expect(
-        attachRepository,
+        repositoryClass,
         isNotNull,
-        reason: 'Missing class named ${testClassName}Repository.',
+        reason: 'Missing class named $repositoryClassName.',
       );
     });
+
+    group('then the $repositoryClassName class', () {
+      group('has a find method', () {
+        var insertRowMethod = CompilationUnitHelpers.tryFindMethodDeclaration(
+          repositoryClass!,
+          name: 'find',
+        );
+        test('defined', () {
+          expect(
+            CompilationUnitHelpers.hasMethodDeclaration(
+              repositoryClass,
+              name: 'find',
+            ),
+            isTrue,
+          );
+        });
+
+        test('that returns a future with a list of the base class', () {
+          expect(
+            insertRowMethod?.returnType?.toSource(),
+            contains('Future'),
+          );
+
+          expect(
+            insertRowMethod?.returnType?.toSource(),
+            contains('List<$testClassName>'),
+          );
+        });
+
+        test('that takes the session as a required param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('Session session'),
+          );
+        });
+
+        test('that takes the where callback as a named optional param', () {
+          var params = insertRowMethod?.parameters?.toSource();
+          expect(
+            params,
+            contains('${testClassName}ExpressionBuilder? where'),
+          );
+        });
+
+        test('that takes the limit int as an optional param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('int? limit'),
+          );
+        });
+
+        test('that takes the offset int as an optional param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('int? offset'),
+          );
+        });
+
+        test('that takes the orderBy column as an optional param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('Column? orderBy'),
+          );
+        });
+
+        test('that takes the orderDescending bool as an optional param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('bool orderDescending'),
+          );
+        });
+
+        test('that takes the orderByList as an optional param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('orderByList'),
+          );
+        });
+
+        test('that takes the transaction object as an optional param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('Transaction? transaction'),
+          );
+        });
+      });
+
+      group('has a findRow method', () {
+        var insertRowMethod = CompilationUnitHelpers.tryFindMethodDeclaration(
+          repositoryClass!,
+          name: 'findRow',
+        );
+        test('defined', () {
+          expect(
+            CompilationUnitHelpers.hasMethodDeclaration(
+              repositoryClass,
+              name: 'findRow',
+            ),
+            isTrue,
+          );
+        });
+
+        test('that returns a future with a nullable base class', () {
+          expect(
+            insertRowMethod?.returnType?.toSource(),
+            contains('Future'),
+          );
+
+          expect(
+            insertRowMethod?.returnType?.toSource(),
+            contains('$testClassName?'),
+          );
+        });
+
+        test('that takes the session as a required param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('Session session'),
+          );
+        });
+
+        test('that takes the where callback as a named optional param', () {
+          var params = insertRowMethod?.parameters?.toSource();
+          expect(
+            params,
+            contains('${testClassName}ExpressionBuilder? where'),
+          );
+        });
+
+        test('that takes the transaction object as an optional param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('Transaction? transaction'),
+          );
+        });
+      });
+
+      group('has a findById method', () {
+        var insertRowMethod = CompilationUnitHelpers.tryFindMethodDeclaration(
+          repositoryClass!,
+          name: 'findById',
+        );
+        test('defined', () {
+          expect(
+            CompilationUnitHelpers.hasMethodDeclaration(
+              repositoryClass,
+              name: 'findById',
+            ),
+            isTrue,
+          );
+        });
+
+        test('that returns a future with a nullable base class', () {
+          expect(
+            insertRowMethod?.returnType?.toSource(),
+            contains('Future'),
+          );
+
+          expect(
+            insertRowMethod?.returnType?.toSource(),
+            contains('$testClassName?'),
+          );
+        });
+
+        test('that takes the session as a required param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('Session session'),
+          );
+        });
+
+        test('that takes an int as a required param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('int id'),
+          );
+        });
+
+        test('that takes the transaction object as an optional param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('Transaction? transaction'),
+          );
+        });
+      });
+
+      group('has an insert row method', () {
+        var insertRowMethod = CompilationUnitHelpers.tryFindMethodDeclaration(
+          repositoryClass!,
+          name: 'insertRow',
+        );
+        test('defined', () {
+          expect(
+            CompilationUnitHelpers.hasMethodDeclaration(
+              repositoryClass,
+              name: 'insertRow',
+            ),
+            isTrue,
+          );
+        });
+
+        test('that returns a future with the base class', () {
+          expect(
+            insertRowMethod?.returnType?.toSource(),
+            contains('Future'),
+          );
+
+          expect(
+            insertRowMethod?.returnType?.toSource(),
+            contains(testClassName),
+          );
+        });
+
+        test('that takes the session as a required param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('Session session'),
+          );
+        });
+
+        test('that takes the row as a required param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('$testClassName row'),
+          );
+        });
+
+        test('that takes the transaction object as an optional param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('Transaction? transaction'),
+          );
+        });
+      });
+
+      group('has an update row method', () {
+        var insertRowMethod = CompilationUnitHelpers.tryFindMethodDeclaration(
+          repositoryClass!,
+          name: 'updateRow',
+        );
+        test('defined', () {
+          expect(
+            CompilationUnitHelpers.hasMethodDeclaration(
+              repositoryClass,
+              name: 'updateRow',
+            ),
+            isTrue,
+          );
+        });
+
+        test('that returns a future with the base class', () {
+          expect(
+            insertRowMethod?.returnType?.toSource(),
+            contains('Future'),
+          );
+
+          expect(
+            insertRowMethod?.returnType?.toSource(),
+            contains(testClassName),
+          );
+        });
+
+        test('that takes the session as a required param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('Session session'),
+          );
+        });
+
+        test('that takes the row as a required param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('$testClassName row'),
+          );
+        });
+
+        test('that takes the transaction object as an optional param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('Transaction? transaction'),
+          );
+        });
+      });
+
+      group('has an delete row method', () {
+        var insertRowMethod = CompilationUnitHelpers.tryFindMethodDeclaration(
+          repositoryClass!,
+          name: 'deleteRow',
+        );
+        test('defined', () {
+          expect(
+            CompilationUnitHelpers.hasMethodDeclaration(
+              repositoryClass,
+              name: 'deleteRow',
+            ),
+            isTrue,
+          );
+        });
+
+        test('that returns a future with the base class', () {
+          expect(
+            insertRowMethod?.returnType?.toSource(),
+            contains('Future'),
+          );
+
+          expect(
+            insertRowMethod?.returnType?.toSource(),
+            contains('int'),
+          );
+        });
+
+        test('that takes the session as a required param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('Session session'),
+          );
+        });
+
+        test('that takes the row as a required param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('$testClassName row'),
+          );
+        });
+
+        test('that takes the transaction object as an optional param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('Transaction? transaction'),
+          );
+        });
+      });
+
+      group('has an delete where method', () {
+        var insertRowMethod = CompilationUnitHelpers.tryFindMethodDeclaration(
+          repositoryClass!,
+          name: 'deleteWhere',
+        );
+        test('defined', () {
+          expect(
+            CompilationUnitHelpers.hasMethodDeclaration(
+              repositoryClass,
+              name: 'deleteWhere',
+            ),
+            isTrue,
+          );
+        });
+
+        test('that returns a future with a list of int', () {
+          expect(
+            insertRowMethod?.returnType?.toSource(),
+            contains('Future'),
+          );
+
+          expect(
+            insertRowMethod?.returnType?.toSource(),
+            contains('List<int>'),
+          );
+        });
+
+        test('that takes the session as a required param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('Session session'),
+          );
+        });
+
+        test('that takes the where callback as a named required param', () {
+          var params = insertRowMethod?.parameters?.toSource();
+          expect(
+            params,
+            contains('required ${testClassName}ExpressionBuilder where'),
+          );
+        });
+
+        test('that takes the transaction object as an optional param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('Transaction? transaction'),
+          );
+        });
+      });
+
+      group('has an count method', () {
+        var insertRowMethod = CompilationUnitHelpers.tryFindMethodDeclaration(
+          repositoryClass!,
+          name: 'count',
+        );
+        test('defined', () {
+          expect(
+            CompilationUnitHelpers.hasMethodDeclaration(
+              repositoryClass,
+              name: 'count',
+            ),
+            isTrue,
+          );
+        });
+
+        test('that returns a future with a list of int', () {
+          expect(
+            insertRowMethod?.returnType?.toSource(),
+            contains('Future'),
+          );
+
+          expect(
+            insertRowMethod?.returnType?.toSource(),
+            contains('int'),
+          );
+        });
+
+        test('that takes the session as a required param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('Session session'),
+          );
+        });
+
+        test('that takes the where callback as a named optional param', () {
+          var params = insertRowMethod?.parameters?.toSource();
+          expect(
+            params,
+            contains('${testClassName}ExpressionBuilder? where'),
+          );
+        });
+
+        test('that takes the limit int as an optional param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('int? limit'),
+          );
+        });
+
+        test('that takes the transaction object as an optional param', () {
+          expect(
+            insertRowMethod?.parameters?.toSource(),
+            contains('Transaction? transaction'),
+          );
+        });
+      });
+    }, skip: repositoryClass == null);
   });
 }
